integreated notification service for all parties

+++ controllers/theatreController.js
@@ -1,13 +1,27 @@
 let theatreModel = require("../models/theatreModel");
 let moviesModel = require("../models/movieModel");
+let sendEmail= require("../utils/notificationClient")
+let userModel = require("../models/userModel");
 
 let createTheatre = async (req, res) => {
 
     try {
         let theatreBody = req.body;
 
+        let client = await userModel.findOne({
+            _id:theatreBody.ownerId
+        })
+
+        if (!client) {
+            return res.status(404).send({
+                message:"No theatre owner found for this ownerid provided"
+            })
+        }
+
         let theatre = await theatreModel.create(theatreBody);
         res.status(200).send(theatre)
+
+        sendEmail(theatre._id,"New theatre created with theatre id: " +theatre._id ,JSON.stringify(theatre),[client.email],"mba-noreply@gmail.com")
 
     } catch (err) {
         console.log(err)
@@ -21,13 +35,39 @@ let updateTheatre = async (req, res) => {
     try {
 
         let updatePassed = req.body;
+        let recepientEmails;
 
+        let requester = await userModel.findOne({
+            userId:req.userId
+        })
+
+       
         let updatedTheatre = await theatreModel.findOneAndUpdate({
             _id: req.params.theatreId
         }, updatePassed, { new: true });
 
+        // sending email to different parties depending upon the request makers
+
+        if (requester.userType == "ADMIN") {
+            let client = await userModel.findOne({
+                _id: updatedTheatre.ownerId
+            });
+            recepientEmails=[client.email]
+        } else {
+            let admins = await userModel.find({
+                userType: "ADMIN",
+                userStatus:"APPROVED"
+            }).select({ email: 1 })
+            
+            admins.map(obj=>recepientEmails.push(obj.email))
+        }
+
+
         if (updatedTheatre) {
             res.status(200).send(updatedTheatre)
+
+            sendEmail(updatedTheatre._id, "Update with theatre id: " + updatedTheatre._id, JSON.stringify(updatedTheatre), [...recepientEmails], "mba-noreply-@gmail.com");
+            
         } else {
             res.status(400).send({
                 message:"No theatre exist with given theatre id"
@@ -108,7 +148,7 @@ let getTheatreById = async (req, res) => {
 
 }
 
-let deleteTheatre = async (req, res) => {
+let deleteTheatre = async (req, res) => { 
     try {
         let isValidTheatreId = await theatreModel.findOne({
                 _id:req.params.theatreId
@@ -119,6 +159,16 @@ let deleteTheatre = async (req, res) => {
             message:"Invalid theatre Id ! no theatre exist with given id"
             })
         }
+
+        let client = await userModel.findOne({
+            _id: isValidTheatreId.ownerId
+        })
+
+        if (!client) {
+            return res.status(404).send({
+                message: "No theatre owner found for this ownerid provided"
+            })
+        }
         await theatreModel.deleteOne({
             _id:req.params.theatreId
         })
@@ -126,6 +176,10 @@ let deleteTheatre = async (req, res) => {
         res.status(200).send({
             message:'successfully removed theatre from the database'
         })
+
+
+        sendEmail(isValidTheatreId._id, "theatre removed with theatre id: " + isValidTheatreId._id +" from the mba-app", JSON.stringify(isValidTheatreId), [client.email], "mba-noreply@gmail.com")
+
 
     } catch (err) {
         console.log(err)
@@ -146,6 +200,12 @@ let addMoviesInTheatre = async (req, res) => {
                 message:"please add the moviesId to be added"
             })
         }
+
+        let recepientEmails;
+
+        let requester = await userModel.findOne({
+            userId: req.userId
+        })
 
         let theatre = await theatreModel.findOne({
             _id: req.params.theatreId
@@ -171,6 +231,22 @@ let addMoviesInTheatre = async (req, res) => {
             }
             
         });
+
+        // sending email to different parties depending upon the request makers
+
+        if (requester.userType == "ADMIN") {
+            let client = await userModel.findOne({
+                _id:theatre.ownerId
+            });
+            recepientEmails = [client.email]
+        } else {
+            let admins = await userModel.find({
+                userType: "ADMIN",
+                userStatus: "APPROVED"
+            }).select({ email: 1 })
+
+            admins.map(obj => recepientEmails.push(obj.email))
+        }
   
         
 
@@ -180,6 +256,8 @@ let addMoviesInTheatre = async (req, res) => {
         res.status(200).send({
             message:"movies added into this theatre"
         })
+
+        sendEmail(theatre._id, "Movies added into theatre id: " + theatre._id, JSON.stringify(theatre), [...recepientEmails], "mba-noreply-@gmail.com");
         
     } catch (err) {
         console.log(err)
@@ -194,6 +272,12 @@ let removeMoviesFromTheatre = async (req, res) => {
     try {
 
         let moviesIdToBeRemoved = req.body.moviesId;
+
+        let recepientEmails;
+
+        let requester = await userModel.findOne({
+            userId: req.userId
+        })
 
         if (moviesIdToBeRemoved == undefined || moviesIdToBeRemoved.length < 1) {
             return res.status(400).send({
@@ -211,6 +295,24 @@ let removeMoviesFromTheatre = async (req, res) => {
                 message:"please pass valid theatre id , no theatre exist for this given id"
             })
         }
+
+
+        // sending email to different parties depending upon the request makers
+
+        if (requester.userType == "ADMIN") {
+            let client = await userModel.findOne({
+                _id: theatre.ownerId
+            });
+            recepientEmails = [client.email]
+        } else {
+            let admins = await userModel.find({
+                userType: "ADMIN",
+                userStatus: "APPROVED"
+            }).select({ email: 1 })
+
+            admins.map(obj => recepientEmails.push(obj.email))
+        }
+
             
         
         theatre.movies = theatre.movies.filter(id => !moviesIdToBeRemoved.includes(id.toString()))
@@ -219,6 +321,8 @@ let removeMoviesFromTheatre = async (req, res) => {
 
 
         res.status(200).send(theatre.movies)
+
+        sendEmail(theatre._id, "Movies removed from theatre id: " + theatre._id, JSON.stringify(theatre), [...recepientEmails], "mba-noreply-@gmail.com");
         
 
     } catch (err) {
@@ -270,6 +374,7 @@ let getAllMoviesInTheatre = async (req, res) => {
 
     }
 }
+
 
 module.exports = { 
     createTheatre,
